requirements.txt

pytest
pytest-qt
PyYAML

guidata==1.7.9
guiqwt==3.0.5
pyqt-tools==1.0.0
PyQt5==5.13.0
pyqt5-plugins==5.15.1.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.1.2
python-dotenv==0.15.0
PythonQwt==0.8.3
pyuic5-tool==0.0.1
qt5-applications==5.15.1.1
QtPy==1.9.0
